[feature/#32] 폴더링
