+++ app/src/main/java/com/teampome/pome/application/PomeApp.kt
@@ -1,4 +1,4 @@
-package com.teampome.pome
+package com.teampome.pome.application
 
 import android.app.Application
 import androidx.appcompat.app.AppCompatDelegate
