+++ app/src/main/java/com/teampome/pome/presentation/friends/screens/FriendsFragment.kt
@@ -1,4 +1,4 @@
-package com.teampome.pome.presentation.friends
+package com.teampome.pome.presentation.friends.screens
 
 import android.os.Bundle
 import android.view.View
@@ -6,9 +6,12 @@ import android.widget.ImageView
 import com.skydoves.balloon.balloon
 import com.teampome.pome.R
 import com.teampome.pome.databinding.FragmentFriendsBinding
-import com.teampome.pome.util.BaseFragment
-import com.teampome.pome.util.FriendsConsumeItemDecorator
-import com.teampome.pome.util.FriendsProfileItemDecorator
+import com.teampome.pome.presentation.friends.*
+import com.teampome.pome.presentation.friends.adapters.FriendsConsumeAdapter
+import com.teampome.pome.presentation.friends.adapters.FriendsProfileAdapter
+import com.teampome.pome.util.base.BaseFragment
+import com.teampome.pome.util.decorate.FriendsConsumeItemDecorator
+import com.teampome.pome.util.decorate.FriendsProfileItemDecorator
 
 class FriendsFragment : BaseFragment<FragmentFriendsBinding>(R.layout.fragment_friends) {
     private lateinit var friendsConsumeAdapter: FriendsConsumeAdapter
