+++ app/src/main/java/com/teampome/pome/presentation/login/screens/AddFriendActivity.kt
@@ -1,9 +1,11 @@
-package com.teampome.pome.presentation.login
+package com.teampome.pome.presentation.login.screens
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.teampome.pome.R
 import com.teampome.pome.databinding.ActivityAddFriendBinding
+import com.teampome.pome.presentation.login.AddFriendAdapter
+import com.teampome.pome.presentation.login.AddFriendData
 
 class AddFriendActivity : AppCompatActivity() {
     private lateinit var binding: ActivityAddFriendBinding
