+++ app/src/main/java/com/teampome/pome/presentation/main/MainActivity.kt
@@ -1,13 +1,13 @@
-package com.teampome.pome
+package com.teampome.pome.presentation.main
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.fragment.app.Fragment
+import com.teampome.pome.R
 import com.teampome.pome.databinding.ActivityMainBinding
-import com.teampome.pome.presentation.friends.FriendsFragment
-import com.teampome.pome.presentation.record.RecordFragment
-import com.teampome.pome.presentation.remind.RemindFragment
-import timber.log.Timber
+import com.teampome.pome.presentation.friends.screens.FriendsFragment
+import com.teampome.pome.presentation.record.screens.RecordFragment
+import com.teampome.pome.presentation.remind.screens.RemindFragment
 
 class MainActivity : AppCompatActivity() {
     private val recordFragment by lazy { RecordFragment() }
