+++ app/src/main/java/com/teampome/pome/presentation/record/screens/RecordLookBackActivity.kt
@@ -1,10 +1,12 @@
-package com.teampome.pome.presentation.record
+package com.teampome.pome.presentation.record.screens
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.teampome.pome.databinding.ActivityRecordLookBackBinding
-import com.teampome.pome.util.CustomItemDecorator
-import com.teampome.pome.util.VerticalItemDecorator
+import com.teampome.pome.presentation.record.RecordAdapter
+import com.teampome.pome.presentation.record.RecordData
+import com.teampome.pome.util.decorate.CustomItemDecorator
+import com.teampome.pome.util.decorate.VerticalItemDecorator
 
 class RecordLookBackActivity : AppCompatActivity() {
