+++ app/src/main/java/com/teampome/pome/util/base/BaseActivity.kt
@@ -1,4 +1,4 @@
-package com.teampome.pome.util
+package com.teampome.pome.util.base
 
 import android.os.Bundle
 import androidx.annotation.LayoutRes
