+++ app/src/main/java/com/teampome/pome/util/decorate/FriendsProfileItemDecorator.kt
@@ -1,4 +1,4 @@
-package com.teampome.pome.util
+package com.teampome.pome.util.decorate
 
 import android.graphics.Rect
 import android.view.View
