+++ app/src/main/res/layout/activity_add_friend.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingHorizontal="16dp"
-    tools:context=".presentation.login.AddFriendActivity">
+    tools:context=".presentation.login.screens.AddFriendActivity">
 
     <TextView
         android:id="@+id/tv_title"
