+++ app/src/main/res/layout/activity_kakao_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_marginHorizontal="16dp"
-    tools:context=".presentation.login.KakaoLoginActivity">
+    tools:context=".presentation.login.screens.KakaoLoginActivity">
 
 
     <ImageView
