+++ app/src/main/res/layout/fragment_no_emotion_dialog.xml
@@ -11,7 +11,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:background="@drawable/inset_dialog_background"
-        tools:context=".presentation.record.NoEmotionDialogFragment">
+        tools:context=".presentation.record.screens.NoEmotionDialogFragment">
 
         <ImageButton
             android:id="@+id/btn_cancel"
