+++ app/src/main/res/layout/fragment_record.xml
@@ -12,7 +12,7 @@
         android:layout_height="match_parent"
         android:background="@color/pome_grey_0"
         android:paddingHorizontal="16dp"
-        tools:context=".presentation.record.RecordFragment">
+        tools:context=".presentation.record.screens.RecordFragment">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/cl_topappbar"
